# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "RoomFinder",
        "description": "RoomFinder API description",
        "contact": {
            "email": "manis6062@gmail.com"
        },
        "version": "1.0.0"
    },
    "host": "127.0.0.1:8000",
    "basePath": "/api/v1",
    "schemes": [
        "http",
        "https"
    ],
    "paths": {
...
        "/users/fblogin": {
            "post": {
                "summary": "Facebook Login",
                "operationId": "fblogin",
                "parameters": [
                    {
                        "name": "device_type",
                        "in": "formData",
                        "description": "Device Type. (ios or android)",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "device_id",
                        "in": "formData",
                        "description": "Device ID.",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "fb_device_token",
                        "in": "formData",
                        "description": "Facebook Access Token.",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "email",
                        "in": "formData",
                        "description": "Email Address",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "profile_pic",
                        "in": "formData",
                        "description": "Profile Pic .",
                        "required": true,
                        "type": "file"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "406": {
                        "description": "not acceptable"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            }
        },
        "/users/logout": {
            "post": {
                "summary": "Users Logout Login",
                "operationId": "logout",
                "parameters": [
                    {
                        "name": "access_token",
                        "in": "header",
                        "description": "Access token",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "406": {
                        "description": "not acceptable"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            }
        },
        "/fav/add-to-favourites": {
            "post": {
                "summary": "Add to favourite",
                "operationId": "AddToFavourite",
                "parameters": [
                    {
                        "name": "access_token",
                        "in": "header",
                        "description": "Access token",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "id",
                        "in": "formData",
                        "description": "User Id",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "id",
                        "in": "formData",
                        "description": "Room Id",
                        "required": false,
                        "type": "integer"
                    },
                    {
                        "name": "id",
                        "in": "formData",
                        "description": "Jagga Id",
                        "required": false,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "406": {
                        "description": "not acceptable"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            }
        }
    },
...
}